improve Java 8 code

+++ src/net/dinkla/raytracer/math/WrappedFloat.java
@@ -27,7 +27,7 @@ public class WrappedFloat implements Comparable<WrappedFloat>{
     }
 
     public float getValue() {
-        return value.floatValue();
+        return value;
     }
     
     public void setValue(final float value) {
